Only serve static files on HEAD or GET requests (#20)
* Only serve static files on HEAD or GET requests

Serving a static file on `POST`, `PUT`, etc. is arguably broken.

We now delegate anything other than `HEAD` or `GET` to upstream.

This enables relevant use cases, e.g. a static log-in page and an
upstream `POST` handler for it.

If upstream doesn't handle a specific method, it should generate an
appropriate status code (currently, ideally 403, as 405 will require
some more work in `wai-middleware-static` to augment any `Allow` header
to be standard compliant).

* Get rid of redundant `mtl` dependency

* Use lts-16.13 for stack builds

* Bump version